+++ packages/commonwealth/client/scripts/views/components/Header/MobileHeader/MobileHeader.scss
