+++ packages/commonwealth/client/scripts/views/components/MobileNavigation/CreateContentDrawer/CreateContentDrawer.scss
