+++ packages/commonwealth/client/scripts/views/components/MobileNavigation/CreateContentDrawer/index.ts
