+++ packages/commonwealth/client/scripts/views/components/MobileNavigation/MobileNavigation.scss
