+++ packages/commonwealth/client/scripts/views/components/MobileNavigation/NavigationButton/NavigationButton.scss
